# smart-backend/docs/swagger.yaml
basePath: /
definitions:
  handlers.APIResponse:
    properties:
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
  handlers.GateActionData:
    properties:
      gate_id:
        example: 1
        type: integer
      status:
        example: true
        type: boolean
    type: object
  handlers.GateActionResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.GateActionData'
      message:
        example: Gate operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
  handlers.GateDTO:
    properties:
      description:
        example: Main vehicle entrance for visitors. Controlled by biometric access,
          opens in 3 seconds with safety sensors.
        type: string
      gate_is_horizontal:
        example: true
        type: boolean
      id:
        example: 1
        type: integer
      is_open:
        example: true
        type: boolean
      location_id:
        example: 1
        type: integer
      title:
        example: Автоматический Шлагбаум №12
        type: string
    type: object
...
  handlers.LocationDTO:
    properties:
      address:
        example: г. Бишкек, проспект Чуй, 135
        type: string
      gates:
        description: Always include gates, even if empty array
        items:
          $ref: '#/definitions/handlers.GateDTO'
        type: array
      id:
        example: 1
        type: integer
      logo:
        example: https://picsum.photos/seed/alatoo/200
        type: string
      title:
        example: Торгово-развлекательный центр Ала-Тоо
        type: string
    type: object
  handlers.LocationsListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/handlers.LocationDTO'
        type: array
      message:
        example: Locations retrieved successfully
        type: string
      success:
        example: true
        type: boolean
    required:
    - message
    - success
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ololo-gate.com
    name: API Support
  description: Secure phone-based authentication backend for Ololo Gate management
    system with dual authentication (users & admins) and role-based access control.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Ololo Gate API
  version: "1.0"
paths:
...
  /api/v1/locations:
    get:
      consumes:
      - application/json
      description: Fetch all locations from third-party API based on user's phone
        with their gates
      produces:
      - application/json
      responses:
        "200":
          description: Locations retrieved successfully
          schema:
            $ref: '#/definitions/handlers.LocationsListResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Get all locations accessible to the current user
      tags:
      - Gate Management
  /api/v1/locations/{gateId}/close:
    put:
      consumes:
      - application/json
      description: Send command to close a specific gate to third-party API
      parameters:
      - description: Gate ID
        in: path
        name: gateId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Gate operation response
          schema:
            $ref: '#/definitions/handlers.GateActionResponse'
        "400":
          description: Invalid gate ID
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Close a gate
      tags:
      - Gate Management
  /api/v1/locations/{gateId}/open:
    put:
      consumes:
      - application/json
      description: Send command to open a specific gate to third-party API
      parameters:
      - description: Gate ID
        in: path
        name: gateId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Gate operation response
          schema:
            $ref: '#/definitions/handlers.GateActionResponse'
        "400":
          description: Invalid gate ID
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "401":
          description: Unauthorized - invalid or missing token
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      security:
      - BearerAuth: []
      summary: Open a gate
      tags:
      - Gate Management
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"